use off_t

+++ include/fm.h
@@ -480,7 +480,7 @@ typedef struct _DownloadList {
 	char *url;
 	char *save;
 	char *lock;
-	size_t size;
+	unsigned long long size;
 	time_t time;
 	int running;
 	int err;
