+++ src/file.c
@@ -119,7 +119,7 @@ loadcmdout(char *cmd,
 #define cur_form_id ((form_sp >= 0)? form_stack[form_sp] : -1)
 	static int form_sp = 0;
 
-	static long long current_content_length;
+	static off_t current_content_length;
 
 	static int cur_hseq;
 #ifdef USE_IMAGE
@@ -6467,7 +6467,7 @@ static char *_size_unit[] = {"b", "kb", "Mb", "Gb", "Tb",
 };
 
 char *
-convert_size(long long size, int usefloat)
+convert_size(off_t size, int usefloat)
 {
 	float csize;
 	int sizepos = 0;
@@ -6483,7 +6483,7 @@ convert_size(long long size, int usefloat)
 }
 
 char *
-convert_size2(long long size1, long long size2, int usefloat)
+convert_size2(off_t size1, off_t size2, int usefloat)
 {
 	char **sizes = _size_unit;
 	float csize, factor = 1;
@@ -6501,7 +6501,7 @@ convert_size2(long long size1, long long size2, int usefloat)
 }
 
 void
-showProgress(long long * linelen, long long * trbyte)
+showProgress(off_t *linelen, off_t *trbyte)
 {
 	int i, j, rate, duration, eta, pos;
 	static time_t last_time, start_time;
@@ -6756,8 +6756,8 @@ void
 loadHTMLstream(URLFile * f, Buffer * newBuf, FILE * src, int internal)
 {
 	struct environment envs[MAX_ENV_LEVEL];
-	long long linelen = 0;
-	long long trbyte = 0;
+	off_t linelen = 0;
+	off_t trbyte = 0;
 	Str lineBuf2 = Strnew();
 #ifdef USE_M17N
 	wc_ces charset = WC_CES_US_ASCII;
@@ -7098,7 +7098,7 @@ loadBuffer(URLFile * uf, Buffer * volatile newBuf)
 	volatile char pre_lbuf = '\0';
 	int nlines;
 	Str tmpf;
-	long long linelen = 0, trbyte = 0;
+	off_t linelen = 0, trbyte = 0;
 	Lineprop *propBuffer = NULL;
 	void(*volatile prevtrap) (int sig) = NULL;
 
@@ -7475,7 +7475,7 @@ getNextPage(Buffer * buf, int plen)
 	Line *volatile top = buf->topLine, *volatile last = buf->lastLine, *volatile cur = buf->currentLine;
 	int i;
 	int volatile nlines = 0;
-	long long linelen = 0, trbyte = buf->trbyte;
+	off_t linelen = 0, trbyte = buf->trbyte;
 	Str lineBuf2;
 	char volatile pre_lbuf = '\0';
 	URLFile uf;
@@ -7592,7 +7592,7 @@ int
 save2tmp(URLFile uf, char *tmpf)
 {
 	FILE *ff;
-	long long linelen = 0, trbyte = 0;
+	off_t linelen = 0, trbyte = 0;
 	Str buf;
 	void(*volatile prevtrap) (int sig) = NULL;
 	static sigjmp_buf env_bak;
@@ -7729,7 +7729,7 @@ _MoveFile(const char *path1, const char *path2)
 	InputStream f1;
 	FILE *f2;
 	int is_pipe;
-	long long linelen = 0, trbyte = 0;
+	off_t linelen = 0, trbyte = 0;
 	Str buf;
 
 	f1 = openIS(path1);
@@ -7770,7 +7770,7 @@ _doFileCopy(char *tmpf, char *defstr, int download)
 	pid_t pid;
 	char *lock;
 	struct stat st;
-	long long size = 0;
+	off_t size = 0;
 	int is_pipe = FALSE;
 
 	if (fmInitialized) {
